feat: update game_mode header with w struct of size 8 to avoid padding problem

+++ include/ecs/component/game_mode_component.h
@@ -8,5 +8,5 @@
 #pragma once
 
 typedef struct game_mode_component {
-    int mode;
+    double mode;
 } c_game_mode_t;
